WebCL-conformance/conformance/extension/WEBCL_html_image/cl_context_createBuffer_with_HTMLImageElement.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLContext createBuffer with HTMLImageElement conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
debug("WebCLContext :: createBuffer with HTMLImageElement test case.");
debug("(CLenum memFlags, HTMLImageElement srcImage)\n");
var image, webCLContext, webCLPlatform, webCLDevices, maxMemAllocSize = 0, invalid_image, imageInvalid;
var wtu = WebCLTestUtils;
webCLContext = wtu.createContext();

try {
    image = new Image();
    image.src = "../../../resources/pebbleHalf.jpg";
    image.onload = execute;
} catch(e) {
    testFailed("Unable to load HTML image element : " + e.name);
}
function execute() {
    try {
        shouldThrowExceptionName("webCLContext.createBuffer(webcl.MEM_READ_ONLY, image);", "WEBCL_EXTENSION_NOT_ENABLED");
        if (!wtu.enableExtension(webcl, "WEBCL_html_image")) {
            throw { description : "Failed to enable extension : WEBCL_html_image. Exitting..."};
        }

        wtu.releaseAll(webCLContext);
        webCLContext = wtu.createContext();

        shouldBeType("webCLContext.createBuffer(webcl.MEM_READ_ONLY, image);", "WebCLBuffer");
        shouldBeType("webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, image);", "WebCLBuffer");
        shouldBeType("webCLContext.createBuffer(webcl.MEM_READ_WRITE, image);", "WebCLBuffer");

        // Testing for exceptions.
        shouldThrowExceptionName("webCLContext.createBuffer(invalid_CLenum, image);", "INVALID_VALUE");
        invalid_image = new Image();
        shouldThrowExceptionName("webCLContext.createBuffer(webcl.MEM_READ_WRITE, invalid_image);", "INVALID_HOST_PTR");

        imageInvalid = image;
        imageInvalid.height = 0;
        shouldThrowExceptionName("webCLContext.createBuffer(webcl.MEM_READ_WRITE, imageInvalid);", "INVALID_IMAGE_SIZE");
        imageInvalid.height = 10;
        imageInvalid.width = 0;
        shouldThrowExceptionName("webCLContext.createBuffer(webcl.MEM_READ_WRITE, imageInvalid);", "INVALID_IMAGE_SIZE");

        webCLPlatform = wtu.getPlatform();
        webCLDevices = wtu.getDevices(webCLPlatform);
        for (i = 0; i < webCLDevices.length; i++) {
            if (maxMemAllocSize < webCLDevices[i].getInfo(webcl.DEVICE_MAX_MEM_ALLOC_SIZE))
                maxMemAllocSize = webCLDevices[i].getInfo(webcl.DEVICE_MAX_MEM_ALLOC_SIZE);
        }
        maxMemAllocSize = webCLDevices[0].getInfo(webcl.DEVICE_MAX_MEM_ALLOC_SIZE);
        maxMemAllocSize *= Int32Array.BYTES_PER_ELEMENT;
        imageInvalid.height = maxMemAllocSize;
        imageInvalid.width = maxMemAllocSize;
        shouldThrowExceptionName("webCLContext.createBuffer(webcl.MEM_READ_WRITE, imageInvalid);", "INVALID_IMAGE_SIZE");

    } catch(e) {
        testFailed(e.description);
    }

    wtu.appendPostJSToHTML(document);
}
successfullyParsed = true;
</script>
</body>
</html>
